+++ include/utils.h
@@ -2,10 +2,10 @@
 #define UTILS_H
 
 // IS FUNCTIONS
-int isKeyword(char * s);
-int isNumber(char * s);
-int isSymbol(char s);
-int isTrimable(char s);
+short int isKeyword(char * s);
+short int isNumber(char * s);
+short int isSymbol(char s);
+short int isTrimable(char s);
 
 char * getErrorCode(int code);
 
@@ -13,7 +13,7 @@ char * getErrorCode(int code);
 char * mallocStr(char *);
 char * sliceStr(char * str, int start, int end);
 char * trimStr(char * str);
-int calcSpaces(char * str);
-int nextQuote(char * str, int index);
+short int calcSpaces(char * str);
+short int nextQuote(char * str, int index);
 
 #endif
